// src/app/api/fn/sales/sales-quotations-list.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { HttpClient, HttpContext, HttpResponse } from '@angular/common/http';
import { Observable } from 'rxjs';
import { filter, map } from 'rxjs/operators';
import { StrictHttpResponse } from '../../strict-http-response';
import { RequestBuilder } from '../../request-builder';

import { PaginatedQuotationList } from '../../models/paginated-quotation-list';

export interface SalesQuotationsList$Params {

/**
 * Customer id.
 */
  customer?: number;

/**
 * On or after this date.
 */
  date_from?: string;

/**
 * On or before this date.
 */
  date_to?: string;

/**
 * Draft, sent or accepted.
 */
  open?: 'false' | 'true';

/**
 * A page number within the paginated result set.
 */
  page?: number;

/**
 * Matches codes, names and numbers.
 */
  search?: string;

/**
 * Document status.
 */
  status?: 'ACCEPTED' | 'DRAFT' | 'INVOICED' | 'REJECTED' | 'SENT';
}

export function salesQuotationsList(http: HttpClient, rootUrl: string, params?: SalesQuotationsList$Params, context?: HttpContext): Observable<StrictHttpResponse<PaginatedQuotationList>> {
  const rb = new RequestBuilder(rootUrl, salesQuotationsList.PATH, 'get');
  if (params) {
    rb.query('customer', params.customer, {});
    rb.query('date_from', params.date_from, {});
    rb.query('date_to', params.date_to, {});
    rb.query('open', params.open, {});
    rb.query('page', params.page, {});
    rb.query('search', params.search, {});
    rb.query('status', params.status, {});
  }

  return http.request(
    rb.build({ responseType: 'json', accept: 'application/json', context })
  ).pipe(
    filter((r: any): r is HttpResponse<any> => r instanceof HttpResponse),
    map((r: HttpResponse<any>) => {
      return r as StrictHttpResponse<PaginatedQuotationList>;
    })
  );
}

salesQuotationsList.PATH = '/api/sales/quotations/';